requirements.txt
# AI-Based Underwater Image Enhancement System - Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Libraries
tensorflow==2.13.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-image==0.21.0

# Database
motor==3.3.2
pymongo==4.5.0

# Image Processing & Metrics
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# API & Utilities
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.0

# Development
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Edge Deployment (optional)
# tensorrt==8.6.1
# pycuda==2022.2.2

# Monitoring
psutil==5.9.5